+++ src/ray/pubsub/publisher.h
@@ -27,7 +27,7 @@
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "absl/synchronization/mutex.h"
-#include "ray/asio/periodical_runner.h"
+#include "ray/asio/periodical_runner_interface.h"
 #include "ray/common/id.h"
 #include "ray/pubsub/publisher_interface.h"
 #include "ray/rpc/rpc_callback_types.h"
